drm/msm/sde: Introduce KCAL color control
Use PCC and HSIC interfaces to manually calibrate the colors of the
DSI panel via sysfs.  This is a simple driver which uses predefined
API to do this. As it is a purely platform driver, sysfs is located
at /sys/devices/platform/kcal_ctrl.0

Supported PCC and HSIC features:
- RGB control,
- Color distortion control,
- Saturation intensity control,
- Contrast intensity control,
- Color brightness control,
- Grayscale mode,
- Astronomy mode;

We already have atomic ops to alter DSI colors, but users prefer
this interface over that.

Change-Id: I4c6f2dd2db5ecba858049790a0a649a28ca68f9c
Signed-off-by: Alesaiko <solcmdr@gmail.com>
Signed-off-by: Panchajanya Sarkar <panchajanya@azure-dev.live>